KeepAliveTask splitted into separate class to remove duplicate code
